requirement.txt
asgiref==3.4.1
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
distro==1.6.0
Django==3.2.9
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.3
numpy==1.21.4
opencv-python==4.5.4.58
openpyxl==3.0.9
pandas==1.3.4
pdfminer.six==20211012
pdftables-api @ git+https://github.com/pdftables/python-pdftables-api.git@a2d329a346197c4dcfc6d1b63b02a89dc87481e6
pycparser==2.21
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tabula-py==2.3.0
tabulate==0.8.9
urllib3==1.26.7
